DL-17282 | Change Identifier type from Int to Long

+++ app/models/pla/response/ReadProtectionsResponse.scala
@@ -37,7 +37,7 @@ object ProtectionRecordsList {
 }
 
 case class ProtectionRecord(
-    identifier: Int,
+    identifier: Long,
     sequenceNumber: Int,
     `type`: ProtectionType,
     certificateDate: String,
